Add dtype argument to model and fitter classes

+++ src/prfmodel/fitters/backend/torch.py
@@ -19,7 +19,7 @@ class TorchSGDFitter(BaseSGDFitter):
 
         params = ParamsDict({v.name: v.value for v in self.trainable_variables + self.non_trainable_variables})
 
-        y_pred = self.model(x, params)
+        y_pred = self.model(x, params, dtype=self.dtype)
 
         loss = self.compute_loss(y=y, y_pred=y_pred)
